requirements.txt
requests~=2.24.0
python-telegram-bot~=12.8
python-dotenv~=0.14.0
gspread~=3.6.0
SQLAlchemy~=1.3.19
pytz~=2020.1
oauth2client~=4.1.3
psycopg2~=2.8.6